Add files via upload
These are the file for SEG2105 assignment 1. I have made changes to EchoServer.java, ClientConsole.java, ChatClient.java, and ServerConsole.java. It also contains a file or the test cases. E5 and E6 was done in class ChatClient, E7 a) was done in ClientConsole, b) and c) was done in ServerConsole.